requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.3
flask-login==0.6.2
flask-wtf==1.1.1
flask-mail==0.9.1
werkzeug==2.2.3
jinja2==3.1.2
sqlalchemy==2.0.9
python-dotenv==1.0.0
email-validator==2.0.0
pandas==2.0.0
numpy==1.24.3
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
openpyxl==3.1.2
xlrd==2.0.1
pillow==9.5.0
fpdf2==2.7.4
gunicorn==20.1.0
google-generativeai